requirements.txt
slackclient>=2.5.0
